Improve build-failure diagnostics and isolate cargo tests
`BinaryLocateError::BuildFailed` now wraps `BuildFailureCapture` whose
`Display` includes lossy-decoded stdout and stderr so logs surface
compiler output, not only the exit status.

Move tests that spawn `cargo` from the library unit-test module into
`tests/binary_test_support_cargo.rs`, omitted on Windows via crate-level
`cfg(not(windows))`, so nested `cargo` is not exercised from the default
unit-test binary on that platform.
